+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Changed
 - parse() now returns a dictionary that includes inferred edges.
+- parse() now returns a dictionary of edges to text.
 - English parser improved.
 
 ### Removed